Make various src/app code compile with -Wconversion (#3681)
We don't actually enable -Wconversion for this code yet, because it's
compiled directly into example apps, so we would need to enable
-Wconversion for those, and they aren't quite there.

+++ src/app/clusters/door-lock-server/door-lock-server.h
@@ -228,7 +228,7 @@ typedef struct
 // space available (spaceAvail) and if so it will send a DefaultResponse
 // command with the status of EMBER_ZCL_STATUS_INSUFFICIENT_SPACE and return
 // false. Otherwise, it will return true.
-bool emAfPluginDoorLockServerCheckForSufficientSpace(uint8_t spaceReq, uint8_t spaceAvail);
+bool emAfPluginDoorLockServerCheckForSufficientSpace(uint16_t spaceReq, uint8_t spaceAvail);
 #endif
 
 // Critical Message Queue
